requirements.txt
# GUI Framework
PyQt5==5.15.10

# Disk Image & File System Forensics
pytsk3==20250801

# Registry Parsing
python-registry==1.3.1

# Utilities
pyyaml==6.0.1

libewf-python==20240506

python-evtx==0.6.1
